Revert "STUTL-30 replace json2csv (#64)" (#65)
This reverts commit b427cee083f1163da4c6eaf7c092f5678cbec732.

Maaaaaaybe this is the cause of failures in other repos, indicating that, across the board, jest needs to be updated to v29, and/or `@json2csv` needs to be added to `jest.config.js::transformIgnorePatterns`. I don't love either of those options, but I guess the "real" problem here is that `@folio` modules aren't transpiled upon distribution and that's allowing this issue to leak through.