Revert "STUTL-33: escapeCqlValue for " \ ^ * ? (#69)" (#74)
This reverts commit 2a773643b91937dd0232657c8473170592493daa.

This appears to be a likely cause of > 150 e2e test failures when tests
rely on an asterisk as part of a wildcard search. We may eventually
decide to keep this work, alongside a function that permits the
asterisk, as outlined in STUTL-37 / PR #73, but we'll want to rearrange the
order of operations (implement that function, then convert existing
search code to use that function, and only then implement the breaking
change outlined in STUTL-33).